Implemented mix function

+++ backend/mixer/levels.hh
@@ -2,8 +2,10 @@
 #define LEVELS_HH
 
 struct Level {
-	int32_t pan;
-	int32_t level;
+	// pan is a float between -1 (all the way to the left) and 1 (all the way to the right)
+	float pan;
+	// level is a float between 0 and 1
+	float level;
 };
 
 #endif
